requirements.txt
# AutoWP-ChinaExport-Bot
# WordPress Automated Content Generation System

requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0
Pillow>=10.1.0
gitpython>=3.1.40
 schedule>=1.2.0
